fix(bash): honor sub-second timeouts (#172)
* fix(bash): honor sub-second timeouts

timeout_ms under 1000 became int(ms/1000)=0, so a documented 500ms
budget timed out immediately. Pass the float seconds through to the session.

* test(bash): lock sub-second timeout via BashTool

Regression must go through BashTool.execute(timeout=500), not only
ShellSession, so int(ms/1000) cannot silently regress.

+++ chapter5/coding-agent/tools/shell_session.py
@@ -57,7 +57,7 @@ class ShellSession:
         self.process = None
         self.start()
     
-    def execute(self, command: str, timeout: int = 120) -> Tuple[str, int]:
+    def execute(self, command: str, timeout: float = 120) -> Tuple[str, int]:
         """Execute command in the persistent shell"""
         self.start()
